requirements.txt
# Core
Django==5.0.6
celery==5.4.0
cryptography==43.0.3
python-dotenv==1.0.1
requests==2.31.0
gunicorn==22.0.0
gevent==24.2.1
drf-yasg==1.21.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.2
django-environ==0.11.2
pika==1.3.2
django-redis==5.4.0
boto3==1.37.13
psycopg2-binary==2.9.9
django_tenants==3.6.1